requirements.txt
dnspython >= 1.16.0